+++ source/CargoSpace/Assets/Scripts/Bus.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 929781179a8eb3347ae3194962b3a9bd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
