+++ source/CargoSpace/Assets/Scripts/PlayerShipMovementBehavior.cs
@@ -2,6 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Bus;
 using Logging;
 using UnityEngine;
 using UnityEngine.Serialization;
