+++ source/CargoSpace/Assets/Scripts/TargetableBehavior.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Bus;
 using Logging;
 using UnityEngine;
